reduce verbosity
the call to open3_and_fdesc3 does not work if placed before the
population of all the I/O API variables.  I'm not sure why it works this
what and not the other way; if seems from the fdesc3 documentation like
the call to fdesc3 should overwrite everything this subroutine is doing
with uninitialized values.