begin implementing whelan/kettle 'hybrid' Fsoil
- implement start date, time and time step calculation for hybrid file
- implement run driver and fortran makefile

Turns out that the "installed" I/O API libraries in $HOME/local/include
is not the most recent version.  The library date is 2012262, while
m3utilio.f source file has a change log entry for March 2014.  This is
relevant because LASTTIME was added in March 2014, so the installed
version does not have it.  This commit has LASTTIME source code pasted
into meld_whelan_kettle_fsoils.F90 to make it available.  The correct
way to solve this is to compile and install the more recent I/O API; I'm
going to try that next.

update: looked into compiling I/O API; the file
ioapi/Makeinclude.Linux2_x86_64pg notes that compiling with Portland
Group's compiler can be problematic because netcdf libraries are often
compiled with gcc/gfortran and this means you have to either (1)
recompile everything (netcdf, etc) with consistent compiler and flags
or (2) sort through all the compiler name mangling to make sure
everybody can talk to one another.  I don't want to take either of these
on right now so I'm going to punt and stick with the bandaid of pasting
in LASTTIME for now.